moved HalResourceContextImpl and HalPathImpl into context package
