+++ comparison/src/main/java/io/wcm/caravan/hal/comparison/impl/HalComparisonImpl.java
@@ -25,9 +25,10 @@ import io.wcm.caravan.hal.comparison.HalComparison;
 import io.wcm.caravan.hal.comparison.HalComparisonStrategy;
 import io.wcm.caravan.hal.comparison.HalComparisonSource;
 import io.wcm.caravan.hal.comparison.HalDifference;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
+import io.wcm.caravan.hal.comparison.impl.context.HalPathImpl;
 import io.wcm.caravan.hal.comparison.impl.embedded.EmbeddedProcessing;
 import io.wcm.caravan.hal.comparison.impl.embedded.EmbeddedProcessingImpl;
-import io.wcm.caravan.hal.comparison.impl.halpath.HalPathImpl;
 import io.wcm.caravan.hal.comparison.impl.links.LinkProcessing;
 import io.wcm.caravan.hal.comparison.impl.links.LinkProcessingImpl;
 import io.wcm.caravan.hal.comparison.impl.properties.PropertyDiffDetector;
