+++ comparison/src/main/java/io/wcm/caravan/hal/comparison/impl/HalComparisonRecursionImpl.java
@@ -27,6 +27,7 @@ import java.util.Set;
 import io.wcm.caravan.hal.comparison.HalComparisonSource;
 import io.wcm.caravan.hal.comparison.HalDifference;
 import io.wcm.caravan.hal.comparison.HalComparisonContext;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.embedded.EmbeddedProcessing;
 import io.wcm.caravan.hal.comparison.impl.links.LinkProcessing;
 import io.wcm.caravan.hal.comparison.impl.properties.PropertyProcessing;
