+++ comparison/src/main/java/io/wcm/caravan/hal/comparison/impl/embedded/EmbeddedProcessingStep.java
@@ -22,7 +22,7 @@ package io.wcm.caravan.hal.comparison.impl.embedded;
 import java.util.List;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.resource.HalResource;
 
 /**
