+++ comparison/src/main/java/io/wcm/caravan/hal/comparison/impl/links/LinkProcessing.java
@@ -19,8 +19,8 @@
  */
 package io.wcm.caravan.hal.comparison.impl.links;
 
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.ProcessingResult;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.resource.HalResource;
 import io.wcm.caravan.hal.resource.Link;
