+++ comparison/src/main/java/io/wcm/caravan/hal/comparison/impl/links/steps/LinkCountMismatchDetector.java
@@ -27,8 +27,8 @@ import java.util.List;
 import com.google.common.collect.ImmutableList;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.HalDifferenceImpl;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.links.LinkProcessingStep;
 import io.wcm.caravan.hal.resource.Link;
