+++ comparison/src/test/java/io/wcm/caravan/hal/comparison/impl/links/steps/AdditionalOrMissingNamedLinkDetectorTest.java
@@ -40,8 +40,8 @@ import org.junit.Test;
 import com.google.common.collect.ImmutableMap;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
-import io.wcm.caravan.hal.comparison.impl.TestHalComparisonContext;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
+import io.wcm.caravan.hal.comparison.testing.TestHalComparisonContext;
 import io.wcm.caravan.hal.resource.Link;
 
 public class AdditionalOrMissingNamedLinkDetectorTest {
