+++ comparison/src/test/java/io/wcm/caravan/hal/comparison/impl/links/steps/LinkTemplateProcessorTest.java
@@ -33,8 +33,8 @@ import org.junit.Before;
 import org.junit.Test;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
-import io.wcm.caravan.hal.comparison.impl.TestHalComparisonContext;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
+import io.wcm.caravan.hal.comparison.testing.TestHalComparisonContext;
 import io.wcm.caravan.hal.resource.Link;
