+++ comparison/src/test/java/io/wcm/caravan/hal/comparison/testing/processing/IgnoreAllEmbeddedResources.java
@@ -23,7 +23,7 @@ import java.util.Collections;
 import java.util.List;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.embedded.EmbeddedProcessingStep;
 import io.wcm.caravan.hal.resource.HalResource;
