+++ comparison/src/test/java/io/wcm/caravan/hal/comparison/testing/processing/ReportAllEmbeddedResources.java
@@ -24,8 +24,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import io.wcm.caravan.hal.comparison.HalDifference;
-import io.wcm.caravan.hal.comparison.impl.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.HalDifferenceImpl;
+import io.wcm.caravan.hal.comparison.impl.context.HalComparisonContextImpl;
 import io.wcm.caravan.hal.comparison.impl.embedded.EmbeddedProcessingStep;
 import io.wcm.caravan.hal.resource.HalResource;
